server/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.10.0
bs4==0.0.1
Django==3.2.3
django-cors-headers==3.7.0
django-seed==0.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Faker==9.8.2
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.1
text-unidecode==1.3
toposort==1.7
